Address review comments: return on error in testutil funcs
