+++ pkg/cmd/webhooks/forward_test.go
@@ -3,10 +3,8 @@ package webhooks
 import (
 	"bytes"
 	"crypto/tls"
-	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"strings"
 	"sync"
 	"testing"
@@ -15,7 +13,6 @@ import (
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/google/shlex"
-	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -74,15 +71,15 @@ func TestForwardRun(t *testing.T) {
 		},
 	}
 
-	wsServer := getWSServer()
+	wsServer := getWSServer(t)
 	defer wsServer.Close()
 
-	ghAPIServer := getGHAPIServer(wsServer.URL)
+	ghAPIServer := getGHAPIServer(wsServer.URL, t)
 	defer ghAPIServer.Close()
 
 	var wg sync.WaitGroup
 	wg.Add(1)
-	webhookRcvServer, forwarded := getWebhookRcvServer(wg.Done)
+	webhookRcvServer, forwarded := getWebhookRcvServer(wg.Done, t)
 	defer webhookRcvServer.Close()
 	webhookRcvServerURL = webhookRcvServer.URL
 
@@ -107,85 +104,4 @@ func TestForwardRun(t *testing.T) {
 	wg.Wait()
 	assert.Equal(t, "lol\n", string(forwarded.event.Body))
 	assert.Equal(t, forwarded.event.Header.Get("Someheader"), "somevalue")
-}
-
-type forwarded struct {
-	event localEvent
-	done  func()
-}
-
-type localEvent struct {
-	Body   []byte `json:"body"`
-	Header http.Header
-}
-
-func (w *forwarded) ServeHTTP(res http.ResponseWriter, req *http.Request) {
-	var event localEvent
-	err := json.NewDecoder(req.Body).Decode(&event)
-	if err != nil {
-		fmt.Errorf("failed to decode request: %s\n", err)
-	}
-	event.Header = http.Header{}
-	for h := range req.Header {
-		event.Header.Add(h, req.Header.Get(h))
-	}
-	w.event = event
-	_, err = res.Write([]byte("OK"))
-	if err != nil {
-		fmt.Errorf("failed to write response: %s\n", err)
-	}
-
-	w.done()
-}
-
-func getWebhookRcvServer(done func()) (*httptest.Server, *forwarded) {
-	s := &forwarded{done: done}
-	return httptest.NewServer(s), s
-}
-
-func getGHAPIServer(wsServerURL string) *httptest.Server {
-	ghAPIHandler := http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		wsURL := strings.Replace(wsServerURL, "http", "ws", 1)
-		ret := createHookResponse{
-			WsURL: wsURL,
-		}
-		err := json.NewEncoder(res).Encode(ret)
-		if err != nil {
-			fmt.Errorf("failed to write response: %s\n", err)
-		}
-	})
-	return httptest.NewTLSServer(ghAPIHandler)
-}
-
-func getWSServer() *httptest.Server {
-	wsHandler := http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		var upgrader = websocket.Upgrader{}
-		c, err := upgrader.Upgrade(res, req, nil)
-		if err != nil {
-			fmt.Errorf("failed to upgrade: %s\n", err)
-		}
-		defer c.Close()
-
-		header := http.Header{}
-		header.Add("Someheader", "somevalue")
-		msg := wsRequest{
-			Header: header,
-			Body:   []byte(`{"body": "bG9sCg=="}`),
-		}
-		send, _ := json.Marshal(msg)
-		err = c.WriteMessage(1, send)
-		if err != nil {
-			fmt.Errorf("failed to write: %s\n", err)
-		}
-		err = c.WriteMessage(websocket.CloseMessage, websocket.FormatCloseMessage(1000, "woops"))
-		if err != nil {
-			fmt.Errorf("failed to write: %s\n", err)
-		}
-	})
-	return httptest.NewServer(wsHandler)
-}
-
-type wsRequest struct {
-	Header http.Header
-	Body   []byte
 }
